+++ plugins/basecontrollers/basecontrollers.h
@@ -131,6 +131,7 @@ If SetDesired is called, only joint values will be set at every timestep leaving
 
     virtual bool SetPath(TrajectoryBaseConstPtr ptraj)
     {
+        boost::mutex::scoped_lock lock(_mutex);
         if( !!ptraj ) {
             // see if at least one point can be sampled, this make it easier to debug bad trajectories
             vector<dReal> v;
@@ -166,6 +167,7 @@ If SetDesired is called, only joint values will be set at every timestep leaving
         if( _bPause ) {
             return;
         }
+        boost::mutex::scoped_lock lock(_mutex);
         TrajectoryBaseConstPtr ptraj = _ptraj; // because of multi-threading setting issues
         if( !!ptraj ) {
             vector<dReal> sampledata;
@@ -348,6 +350,7 @@ private:
     CollisionReportPtr _report;
     UserDataPtr _cblimits;
     ConfigurationSpecification _samplespec;
+    boost::mutex _mutex;
 };
 
 class RedirectController : public ControllerBase